requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anthropic==0.35.0
anyio==4.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.33.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
docker==7.1.0
fastapi==0.115.0
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langchain==0.3.2
langchain-anthropic==0.2.3
langchain-cohere==0.3.0
langchain-community==0.3.1
langchain-core==0.3.9
langchain-experimental==0.3.2
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langsmith==0.1.131
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.16.2
mlflow-skinny==2.16.2
mock==4.0.3
multidict==6.1.0
multiprocess==0.70.17
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
starlette==0.38.6
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.13.1
yfinance==0.2.44
zipp==3.20.2
